Path validation in the UI now works.
