+++ src/main/java/edu/ucsb/cs56/projects/games/pacman/AIPlayerAvoider.java
@@ -10,8 +10,6 @@ import java.util.Random;
 import java.util.TreeMap;
 
 import edu.ucsb.cs56.projects.games.pacman.GridWalker.Direction;
-import edu.ucsb.cs56.projects.games.pacman.GridWalker.Path;
-import edu.ucsb.cs56.projects.games.pacman.GridWalker.PathSection;
 
 /**
  * This one tries to avoid all the ghosts by running away from the closest
@@ -176,7 +174,7 @@ public class AIPlayerAvoider extends AIPlayer {
 	private Direction getRandomAwayDirection(GridWalker gridWalker) {
 
 		Direction newDirection = null;
-		HashSet<PathSection> ps = gridWalker.getPossiblePaths(gridWalker.new Point(myX, myY));
+		HashSet<PathSection> ps = gridWalker.getPossiblePaths(new Point(myX, myY));
 
 		if (ps != null) {
 			ArrayList<Direction> possibleDirections = new ArrayList<Direction>();
