+++ src/main/java/edu/ucsb/cs56/projects/games/pacman/Board.java
@@ -11,7 +11,6 @@ import edu.ucsb.cs56.projects.games.pacman.Character.PlayerType;
 import edu.ucsb.cs56.projects.games.pacman.DataEvent.DataEventType;
 import edu.ucsb.cs56.projects.games.pacman.GridWalker.Direction;
 import edu.ucsb.cs56.projects.games.pacman.GridWalker.DirectionDistance;
-import edu.ucsb.cs56.projects.games.pacman.GridWalker.Path;
 
 /**
  * Playing field for a Pac-Man arcade game remake that keeps track of all
@@ -587,5 +586,9 @@ public class Board implements Runnable, EventTrackable {
 
 	public int getScore() {
 		return score;
+	}
+	public Path getShortestPath(int x1, int y1, int x2, int y2) {
+		Path path = getGrid().getGridWalker().getShortestPath(x1,  y1,  x2,  y2);
+		return path;
 	}
 }
